requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
numpy
pandas
matplotlib
scikit-learn
pyyaml
